secured product and user routes
