+++ server/api/product.js
@@ -1,5 +1,5 @@
 const router = require('express').Router()
-
+const isAdmin = require('./isAdmin')
 const {Product, User, OrderProduct, Order} = require('../db/models')
 
 router.get('/', async (req, res, next) => {
@@ -22,15 +22,6 @@ router.get('/:productId', async (req, res, next) => {
     next(err)
   }
 })
-
-const isAdmin = (req, res, next) => {
-  if (!req.user.isAdmin) {
-    const err = new Error('No privileges')
-    err.status = 401
-    return next(err)
-  }
-  next()
-}
 
 router.post('/', isAdmin, async (req, res, next) => {
   try {
